fix: rename .wasm to -wasm to appease jest
